refactor: refactor style for detailpage
# 10
